+++ src/pages/settings/InitialSettingsPageView/index.native.tsx
@@ -1,22 +1,10 @@
 import React, {useRef} from 'react';
-// eslint-disable-next-line no-restricted-imports
-import type {NativeScrollEvent, NativeSyntheticEvent, ScrollView as RNScrollView} from 'react-native';
 import ConfirmModal from '@components/ConfirmModal';
 import ScreenWrapper from '@components/ScreenWrapper';
 import ScrollView from '@components/ScrollView';
 import useLocalize from '@hooks/useLocalize';
 import useThemeStyles from '@hooks/useThemeStyles';
-
-type InitialSettingsPageViewProps = {
-    headerContent: React.JSX.Element;
-    scrollViewRef: React.RefObject<RNScrollView | null>;
-    onScroll: (event: NativeSyntheticEvent<NativeScrollEvent>) => void;
-    accountMenuItems: React.JSX.Element;
-    generalMenuItems: React.JSX.Element;
-    shouldShowSignoutConfirmModal: boolean;
-    toggleSignoutConfirmModal: (value: boolean) => void;
-    signOut: (shouldForceSignout?: boolean) => Promise<void> | undefined;
-};
+import type InitialSettingsPageViewProps from './types';
 
 function InitialSettingsPageView({
     headerContent,
