+++ src/pages/settings/InitialSettingsPageView/index.tsx
@@ -1,6 +1,4 @@
 import React, {useRef} from 'react';
-// eslint-disable-next-line no-restricted-imports
-import type {NativeScrollEvent, NativeSyntheticEvent, ScrollView as RNScrollView} from 'react-native';
 import ConfirmModal from '@components/ConfirmModal';
 import NavigationTabBar from '@components/Navigation/NavigationTabBar';
 import NAVIGATION_TABS from '@components/Navigation/NavigationTabBar/NAVIGATION_TABS';
@@ -9,17 +7,7 @@ import ScrollView from '@components/ScrollView';
 import useLocalize from '@hooks/useLocalize';
 import useResponsiveLayout from '@hooks/useResponsiveLayout';
 import useThemeStyles from '@hooks/useThemeStyles';
-
-type InitialSettingsPageViewProps = {
-    headerContent: React.JSX.Element;
-    scrollViewRef: React.RefObject<RNScrollView | null>;
-    onScroll: (event: NativeSyntheticEvent<NativeScrollEvent>) => void;
-    accountMenuItems: React.JSX.Element;
-    generalMenuItems: React.JSX.Element;
-    shouldShowSignoutConfirmModal: boolean;
-    toggleSignoutConfirmModal: (value: boolean) => void;
-    signOut: (shouldForceSignout?: boolean) => Promise<void> | undefined;
-};
+import type InitialSettingsPageViewProps from './types';
 
 function InitialSettingsPageView({
     headerContent,
